refactor upload image closure to controller
