+++ src/controllers/services.js
@@ -1,4 +1,8 @@
 const Service = require("../models/service");
+const { uploadFile, getFileStream } = require("../helpers/s3");
+const fs = require("fs");
+const util = require("util");
+const unlinkFile = util.promisify(fs.unlink);
 
 const createService = async (req, res) => {
   const service = new Service(req.body);
@@ -69,9 +73,28 @@ const updateService = async (req, res) => {
   }
 }
 
+const updateServiceImage = 
+async (req, res) => {
+  //console.log(req)
+  const file = req.file
+  const result = await uploadFile(file);
+  //console.log(result)
+  await unlinkFile(file.path);
+  const service = await Service.findOne({ _id: req.params.id });
+
+    if (!service) {
+      return res.status(404).send();
+    }
+  
+  service["picture_link"] = result.Key
+  await service.save()
+  res.send(service);
+}
+
 module.exports = {
   createService,
   getAllServices,
   getServiceById,
-  updateService
+  updateService,
+  updateServiceImage
 }
